Refactored, optimize implementation. Rename file names
 1. @MainWindow.cs
    From MainWindow, now we DON'T have a RobotCommunicator as a member variable.
   The only way to communicate with robot is to use RobotController.
   When we pressed 'Connect' button, just invoke AddCommunicatorAsync() from RobotController, and give url address and port number as parameters.
   Therefore, all control command should be issued by using RobotController.

 2. @ALL
 Removed unreferenced namespace

 3. @MainWindowExtension.cs
Rename some async methods (it was not suffixed by 'Async'.

 4. @RobotController.cs
Modified comments of class RobotController.

 5. @RobotCommnicator.cs
There was protected version of connectToURLAsync(), but it is proven to be useless, so I removed it. User just invoke ConnectToURLAsync() public method.

 6. @RobotCommunicator.cs
Property 'Connected' is modifed to handle the case where 'ClientSocket' is null.